Add files via upload
This is the very first app I made. It took some time as I didn't want to solve any problems by resorting to an online tutorial - I had learned the relevant material it just needed to be applied correctly. I enjoyed this a lot and began to think about the quality of code in the program (even if insignificant) hence the use of the switch statement instead of multiple ifs.