Add files via upload
This was the second app I built. I built this after searching for projects and coming across a tutorial - however, again, I chose to try and replicate the finished project at the end of the tutorial as opposed to following it through. 
This project was useful to improve my understanding of how JS can interact with CSS for a better UX.